# packages/backend/src/swagger.yaml
openapi: 3.0.0
info:
  title: Vistra Document Management System API
  version: 1.0.0
  description: API for managing documents and folders
servers:
  - url: http://localhost:3001
    description: Development server
paths:
  /:
    get:
      summary: Retrieve a list of documents and folders
      parameters:
        - name: search
          in: query
          description: Search query
          schema:
            type: string
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          description: Number of items per page
          schema:
            type: integer
            default: 25
        - name: sortField
          in: query
          description: Field to sort by
          schema:
            type: string
            default: name
        - name: sortOrder
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        '200':
          description: A list of items
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                  total:
                    type: integer
        '500':
          description: Internal server error
  /folders:
    post:
      summary: Create a new folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - created_by
              properties:
                name:
                  type: string
                parent_id:
                  type: integer
                created_by:
                  type: string
      responses:
        '201':
          description: Folder created
        '400':
          description: Bad request
        '500':
          description: Internal server error
  /folders/{id}:
    put:
      summary: Update a folder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                parent_id:
                  type: integer
      responses:
        '200':
          description: Folder updated
        '400':
          description: Bad request
        '500':
          description: Internal server error
    delete:
      summary: Delete a folder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Folder deleted
        '400':
          description: Bad request
        '500':
          description: Internal server error
  /documents:
    post:
      summary: Create a new document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - created_by
              properties:
                name:
                  type: string
                folder_id:
                  type: integer
                created_by:
                  type: string
                file_size_bytes:
                  type: integer
      responses:
        '201':
          description: Document created
        '400':
          description: Bad request
        '500':
          description: Internal server error
  /documents/{id}:
    put:
      summary: Update a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                folder_id:
                  type: integer
      responses:
        '200':
          description: Document updated
        '400':
          description: Bad request
        '500':
          description: Internal server error
    delete:
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Document deleted
        '400':
          description: Bad request
        '500':
          description: Internal server error